test: make tests that require binaries in path CI only (#1393)
* test: make tests that require binaries in path CI only

* test: break out integration tests

* fix: remove flag in ci tests

* lint: fmt

* fix: use workspace dep
