+++ crates/node-bindings/Cargo.toml
@@ -28,3 +28,6 @@ tempfile.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 url.workspace = true
+
+[dev-dependencies]
+ci_info.workspace = true
